+++ packages/cli/src/commands/addLoggerOptions.ts
@@ -2,9 +2,10 @@ import type { Command } from "commander";
 import { Option } from "commander";
 
 export function addLoggerOptions(program: Command) {
+  const isCI = process.env.CI || process.env.TF_BUILD;
   return program
     .option("--reporter <reporter...>", "reporter", "npmLog")
-    .addOption(new Option("--progress").conflicts("--reporter"))
+    .addOption(new Option("--progress").conflicts("--reporter").default(!isCI))
     .addOption(new Option("--log-level <level>", "log level").choices(["info", "warn", "error", "verbose", "silly"]).conflicts("--verbose"))
     .option("--verbose", "verbose output");
 }