+++ packages/cli/src/commands/run/index.ts
@@ -1,5 +1,4 @@
 import { Command } from "commander";
-import os from "os";
 import { action } from "./action.js";
 import { addLoggerOptions } from "../addLoggerOptions.js";
 import { isRunningFromCI } from "../isRunningFromCI.js";
@@ -8,18 +7,7 @@ const runCommand = new Command("run");
 
 addLoggerOptions(runCommand)
   .action(action)
-  .option(
-    "-c, --concurrency <n>",
-    "concurrency",
-    (value) => {
-      if (value.endsWith("%")) {
-        return (parseInt(value.slice(0, -1)) / 100) * os.cpus().length;
-      } else {
-        return parseInt(value) || os.cpus().length - 1;
-      }
-    },
-    os.cpus().length - 1
-  )
+  .option("-c, --concurrency <n>", "concurrency", (value) => parseInt(value, 10))
   .option("--max-workers-per-task <maxWorkersPerTarget...>", "set max worker per task, e.g. --max-workers-per-task build=2 test=4", [])
   // Common Options
   .option("--scope <scope...>", "scopes the run to a subset of packages (by default, includes the dependencies and dependents as well)")
@@ -27,6 +15,7 @@ addLoggerOptions(runCommand)
   .option("--include-dependencies|--dependencies", 'adds the scoped packages dependencies as the "entry points" for the target graph run')
   .option("--since <since>", "only runs packages that have changed since the given commit, tag, or branch")
   .option("--to <scope...>", "runs up to a package (shorthand for --scope=<scope...> --no-dependents)")
+  .option("--allow-no-target-runs")
 
   // Run Command Options
   .option("--grouped", "groups the logs", false)