+++ packages/cli/src/config/getConfig.ts
@@ -31,6 +31,8 @@ export async function getConfig(cwd: string): Promise<ConfigOptions> {
     ],
     loggerOptions: config?.loggerOptions ?? {},
     runners: config?.runners ?? {},
-    workerIdleMemoryLimit: config?.workerIdleMemoryLimit ?? os.totalmem(), // 0 means no limit
+    workerIdleMemoryLimit: config?.workerIdleMemoryLimit ?? os.totalmem(), // 0 means no limit,
+    concurrency: config?.concurrency ?? os.cpus().length - 1,
+    allowNoTargetRuns: config?.allowNoTargetRuns ?? false,
   };
 }