+++ packages/cli/src/types/ConfigOptions.ts
@@ -55,4 +55,14 @@ export interface ConfigOptions {
    * Maximum worker idle memory, this would cause workers to restart if they exceed this limit. This is useful to prevent memory leaks.
    */
   workerIdleMemoryLimit: number; // in bytes
+
+  /**
+   * Maximum number of concurrent tasks to run
+   */
+  concurrency: number;
+
+  /**
+   * Allows for no targets run
+   */
+  allowNoTargetRuns: boolean;
 }