+++ packages/cli/CHANGELOG.md
@@ -1,8 +1,26 @@
 # Change Log - @lage-run/cli
 
-This log was last generated on Tue, 06 Dec 2022 00:48:02 GMT and should not be manually modified.
+This log was last generated on Thu, 08 Dec 2022 00:49:28 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## 0.5.6
+
+Thu, 08 Dec 2022 00:49:28 GMT
+
+### Patches
+
+- sets the default to progress (kchau@microsoft.com)
+- Bump @lage-run/scheduler to v0.7.2
+- Bump @lage-run/reporters to v0.2.43
+
+## 0.5.5
+
+Tue, 06 Dec 2022 23:28:12 GMT
+
+### Patches
+
+- Bump @lage-run/reporters to v0.2.42
 
 ## 0.5.4