+++ packages/lage2/CHANGELOG.md
@@ -1,8 +1,25 @@
 # Change Log - @lage-run/lage
 
-This log was last generated on Tue, 06 Dec 2022 00:48:02 GMT and should not be manually modified.
+This log was last generated on Thu, 08 Dec 2022 00:49:28 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## 2.5.23
+
+Thu, 08 Dec 2022 00:49:28 GMT
+
+### Patches
+
+- Bump @lage-run/cli to v0.5.6
+- Bump @lage-run/scheduler to v0.7.2
+
+## 2.5.22
+
+Tue, 06 Dec 2022 23:28:12 GMT
+
+### Patches
+
+- Bump @lage-run/cli to v0.5.5
 
 ## 2.5.21